# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DevLens API
  version: 1.0.0
  description: |
    API contract สำหรับ DevLens Engineering Intelligence Platform

    Related documents:
    - 01-business-logic-architecture.md
    - 03-backend-roadmap.md
    - 04-metric-definitions.md
    - 05-database-design.dbml
    - 06-api-design.md
servers:
  - url: /api/v1
    description: Current API version
...
security:
  - bearerAuth: []
paths:
...
  /organizations:
    post:
      tags:
        - Organizations
      summary: Create organization
      description: Creates an organization and assigns the authenticated user as the initial owner.
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
    get:
      tags:
        - Organizations
      summary: Get organizations accessible by current user
      operationId: listOrganizations
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Organization list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationListResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /organizations/{organizationId}:
    get:
      tags:
        - Organizations
      summary: Get organization detail
      operationId: getOrganization
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "200":
          description: Organization detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags:
        - Organizations
      summary: Update organization
      operationId: updateOrganization
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags:
        - Organizations
      summary: Soft delete organization
      operationId: deleteOrganization
      parameters:
        - $ref: "#/components/parameters/OrganizationId"
      responses:
        "204":
          description: Organization deleted
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /repositories/{repositoryId}/dashboard/summary:
    get:
      tags:
        - Dashboard
      summary: Get repository dashboard summary
      operationId: getDashboardSummary
      parameters:
        - $ref: "#/components/parameters/RepositoryId"
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummaryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /repositories/{repositoryId}/dashboard/pr-cycle-time:
    get:
      tags:
        - Dashboard
      summary: Get dashboard PR cycle time metrics
      operationId: getDashboardPullRequestCycleTime
      parameters:
        - $ref: "#/components/parameters/RepositoryId"
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/Interval"
      responses:
        "200":
          description: Pull request cycle time metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullRequestMetricsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /repositories/{repositoryId}/dashboard/review-wait-time:
    get:
      tags:
        - Dashboard
      summary: Get dashboard review wait time metrics
      operationId: getDashboardReviewWaitTime
      parameters:
        - $ref: "#/components/parameters/RepositoryId"
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/Interval"
      responses:
        "200":
          description: Review wait time metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewMetricsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /repositories/{repositoryId}/dashboard/review-queue:
    get:
      tags:
        - Dashboard
      summary: Get dashboard review queue
      operationId: getDashboardReviewQueue
      parameters:
        - $ref: "#/components/parameters/RepositoryId"
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Review queue items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewQueueResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RepositoryId:
      name: repositoryId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    DateFrom:
      name: from
      in: query
      required: true
      description: Inclusive start date
      schema:
        type: string
        format: date
    DateTo:
      name: to
      in: query
      required: true
      description: Inclusive end date
      schema:
        type: string
        format: date
    Interval:
      name: interval
      in: query
      schema:
        type: string
        enum:
          - day
          - week
          - month
        default: day
...
  responses:
    Unauthorized:
      description: Authentication is required or token is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: User does not have permission to access the resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with current resource state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
  schemas:
...
    PaginationMeta:
      type: object
      required:
        - page
        - pageSize
        - totalItems
        - totalPages
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 42
        totalPages:
          type: integer
          example: 3
    Organization:
      type: object
      required:
        - id
        - githubId
        - slug
        - name
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        githubId:
          type: integer
          format: int64
        slug:
          type: string
          example: devlens
        name:
          type: string
          example: DevLens
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          nullable: true
    CreateOrganizationRequest:
      type: object
      required:
        - githubId
        - slug
        - name
      properties:
        githubId:
          type: integer
          format: int64
          example: 123456
        slug:
          type: string
          example: devlens
        name:
          type: string
          example: DevLens
    UpdateOrganizationRequest:
      type: object
      properties:
        slug:
          type: string
          example: devlens-platform
        name:
          type: string
          example: DevLens Platform
    OrganizationResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Organization"
    OrganizationListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
...
    DashboardSummary:
      type: object
      required:
        - repositoryId
        - from
        - to
        - prCycleTimeMinutes
        - reviewWaitMinutes
        - deploymentFrequency
        - changeFailureRate
        - reviewCoverage
      properties:
        repositoryId:
          type: string
          format: uuid
        from:
          type: string
          format: date
        to:
          type: string
          format: date
        prCycleTimeMinutes:
          type: number
          format: double
        reviewWaitMinutes:
          type: number
          format: double
        deploymentFrequency:
          type: number
          format: double
        changeFailureRate:
          type: number
          format: double
          minimum: 0
          maximum: 1
        reviewCoverage:
          type: number
          format: double
          minimum: 0
          maximum: 1
    DashboardSummaryResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/DashboardSummary"
    MetricPoint:
      type: object
      required:
        - date
        - value
      properties:
        date:
          type: string
          format: date
        value:
          type: number
          format: double
    PullRequestMetrics:
      type: object
      required:
        - averageCycleTimeMinutes
        - averageFilesChanged
        - averageAdditions
        - averageDeletions
        - cycleTimeTrend
      properties:
        averageCycleTimeMinutes:
          type: number
          format: double
        averageFilesChanged:
          type: number
          format: double
        averageAdditions:
          type: number
          format: double
        averageDeletions:
          type: number
          format: double
        cycleTimeTrend:
          type: array
          items:
            $ref: "#/components/schemas/MetricPoint"
    PullRequestMetricsResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/PullRequestMetrics"
    ReviewMetrics:
      type: object
      required:
        - averageWaitMinutes
        - averageReviewMinutes
        - reviewCoverage
        - waitTimeTrend
      properties:
        averageWaitMinutes:
          type: number
          format: double
        averageReviewMinutes:
          type: number
          format: double
        reviewCoverage:
          type: number
          format: double
          minimum: 0
          maximum: 1
        waitTimeTrend:
          type: array
          items:
            $ref: "#/components/schemas/MetricPoint"
    ReviewMetricsResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/ReviewMetrics"
...
    ReviewQueueItem:
      type: object
      required:
        - pullRequestId
        - number
        - title
        - author
        - waitingMinutes
      properties:
        pullRequestId:
          type: string
          format: uuid
        number:
          type: integer
        title:
          type: string
        author:
          type: string
        reviewRequestedAt:
          type: string
          format: date-time
          nullable: true
        waitingMinutes:
          type: number
    ReviewQueueResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ReviewQueueItem"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    ErrorDetail:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: REPOSITORY_NOT_FOUND
        message:
          type: string
          example: Repository not found
        requestId:
          type: string
          example: req_01JABCXYZ
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ValidationIssue:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: from
        message:
          type: string
          example: from must be before or equal to to
    ValidationErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          allOf:
            - $ref: "#/components/schemas/ErrorDetail"
            - type: object
              properties:
                details:
                  type: array
                  items:
                    $ref: "#/components/schemas/ValidationIssue"